requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
python-nmap==0.7.1
scapy==2.5.0
dnspython==2.4.2
python-whois==0.8.0
pyOpenSSL==23.3.0
cryptography==41.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0
